+++ src/main/kotlin/mine/sweeper/domain/MineField.kt
@@ -1,6 +1,7 @@
 package mine.sweeper.domain
 
 import mine.sweeper.domain.value.GameStatus
+import mine.sweeper.domain.value.Position
 
 class MineField(position: Position) : Field(position) {
     override fun open(): GameStatus {
