+++ src/main/kotlin/mine/sweeper/domain/SafeField.kt
@@ -1,6 +1,7 @@
 package mine.sweeper.domain
 
 import mine.sweeper.domain.value.GameStatus
+import mine.sweeper.domain.value.Position
 
 class SafeField(position: Position) : Field(position) {
     var value = 0
