+++ src/main/kotlin/mine/sweeper/domain/value/Position.kt
@@ -1,3 +1,3 @@
-package mine.sweeper.domain
+package mine.sweeper.domain.value
 
 data class Position(val x: Int, val y: Int)
