+++ src/main/kotlin/mine/sweeper/view/InputView.kt
@@ -1,8 +1,8 @@
 package mine.sweeper.view
 
-import mine.sweeper.domain.Position
 import mine.sweeper.domain.value.Height
 import mine.sweeper.domain.value.MineCount
+import mine.sweeper.domain.value.Position
 import mine.sweeper.domain.value.Width
 
 object InputView {
